requirements.txt
fastapi
uvicorn
pydantic
cassandra-driver
pydantic_core
